[#9] Update Python tests, fix for bug in plugin code
This PR focuses on updating the Python tests to perform appropriate
cleanup and avoid the use of outdated icommands. Additional changes
include extra logging and fixes for Python syntax warnings.

This PR also fixes an incompatibility in the C++ code found during
testing. This fix updates how parameters are substituted to not
include a trailing comma in a genquery statement.